refactor: types폴더 index를 common으로 변경
